OMAP4: Program a per C-state timeout value for the one-shot timer
While there is CPU load preventing deeper C-states, set a one-shot
timer with a per C-state timeout in order to reevaluate c-state at
its expiration, therefore giving another opportunity to pick a
deeper c-state.

The timeout parameter should be characterized depending on the
platforms and on the power/performance requirement.
Long timeouts values means more time in shallowest c-states.
Short timeouts means more time spent in deepest c-state.

By default on our OMAP platform it was set to the corresponding
c-state target-residency:
- From performance perspective: As the system is busy, do not go
into deepest c-state at least during a period of target-residency.
If after this period, the system is still in idle lets give him
another chance to pick a deeper c-state.
- From power perspective: When picking a c-state we should stay
in it at least the target-residency time to be profitable in
terms of power.

C1 C-state timeout was set to 100us for performance reason;
in order to give the system some time in shallow C-state for
further activities that could potentially start.

Change-Id: Ic861c5f9975fe864bd1e6eba64585a7dc3fd1905
Signed-off-by: Nicole Chalhoub <n-chalhoub@ti.com>
Signed-off-by: Kevin Hilman <khilman@ti.com>