[WA] Set L3_1 and L3_2 in NO_SLEEP during AV playback/record
When L3_1 and L3_2 are set to HW_AUTO during AV playback usecases, hang issue
is seen. When L3_1 and L3_2 are set to NO_SLEEP no more hang.
As a WA put L3_1 and L3_2 to NO_SLEEP when syslink get IVAHD and back to
HW_AUTO when syslink release IVAHD

Change-Id: Id75e1dcc3193ae9d2c507ff6e527566a0109badf
Signed-off-by: Nicole Chalhoub <n-chalhoub@ti.com>