ASoc: twl6040: SW WA for twl6040 DC offset issue
On TWL6040 device, previous version to 1.3,
a DC offset may be observed in below configurations :
• EAR in using SINC Filter
• HS in mono-mode (Right or Left)

The offset is linked to an error in the SINC after reset.

To ensure a correct initial state of the SINCs, the DAC mode
and enable signal need the following requirements:
- The mode of the DAC which is enable/disable at 1st, needs to
perform when the other is still enabled.
- The DAC mode change on the fly is not supported.

The SW WA performs the following sequencing:
- Enable HS DAC Right: Set HSRCTL[0] HSDACENAR to 0
- Set the mode of HS DAC Left: Set HSLCTL[1] HSDACMODEL
- Disable HS DAC Right: Set HSRCTL[0] HSDACENAR to 1
- Enable HS DAC Left: Set HSLCTL[0] HSDACENAL to 0
- Set the mode of HS DAC Right: Set HSRCTL[1] HSDACMODER
- Disable HS DAC Left: Set HSLCTL[0] HSDACENAL to 1

Change-Id: I3f7d4b0f3694f53f837a336ce022932a9e592263
Signed-off-by: Francois Mazard <f-mazard@ti.com>
Signed-off-by: Pierre Moos <p-moos@ti.com>