GPIO: OMAP: fix setting IRQWAKEN bits for OMAP4
Setting the IRQWAKEN bit was overwriting previous IRQWAKEN bits,
causing only the last bit set to take effect, resulting in lost
wakeups when the GPIO controller is in idle.

Replace direct writes to IRQWAKEN with MOD_REG_BIT calls to
perform a read-modify-write on the register.

This is a port of patch from 3.0 branch.

Original-Author: Colin Cross <ccross@android.com>
Original-Author: Kevin Hilman <khilman@ti.com>

Change-Id: I5367247dcabba02082fc536beddd1e3bf504079c
Signed-off-by: Mohan V <mohanv@ti.com>