OMAPLFB: Make sure pending compositions are dropped
NV12 buffers are marked as busy as soon as they are added
to the list of pending compositions. If a creating a
composition fails for whatever reason we want to make
sure those buffers are not busy.

Also dynamically allocate a work_struct for the clone
handler in case it takes too long (with verbose debug)
otherwise it could be reused.

Change-Id: I27b5732db33870b70008ad853aa58f444f996f73
Signed-off-by: Tony Lofthouse <a0741364@ti.com>