OMAPLFB: Clear pending compositions when disabling clone
When playing a video with HDMI attached, if you disconnect
and reconnect the TV you will get an out of order video frame
posted on the TV (the LCD panel is unaffected).

Clear the pending composition queue for the cloned manager.

Also in this patch we free the cloning buffers until after the
the workqueues are flushed and destroyed.

Change-Id: I51edb3405f0c258eb0e8f56cbbd0f2d2d639f3cd
Signed-off-by: Tony Lofthouse <a0741364@ti.com>