OMAP4430: Thermal: Duty Cycle: Fix heating_budget burst consumption
In case of consumption of the heating_budget in bursts smaller than
the heating time, the current code is not behaving as expected.

The consumption of the heating_budget is done wrongly, mainly because
the decrements are being done in jiffies, instead of in miliseconds.

A part from that, the way the burst size is currently computed may
include the overhead of scheduling the workqueue. Meaning, the decrements
on the heating_budget could be greater than what the burst could really be.

This patch is changing the way the decrements are computed. Now,
every time we see a frequency change to the maximum frequency, we mark
the time we entered the HEAT state. When we exit the maximum frequency
before the heating budget has elapsed, before we enter COOL1, we decrease
from the budget the burst based on the mark we used.

In case jiffies is wrapping, we are currently considering that
heating_budget has elapsed.

Also this patch will remove the unnecessary 1ms delay while queueing
the heating workqueue.

Change-Id: I07cd724f5381ee0f1a3608ab05476fc3017e1dac
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>