SGX: OMAPLFB: Fix lock screen not being shown on resume
When suspend happens the ProcessFlip requests in OMAPLFB are
ignored and the create compositions requests too as well.
This behavior makes the ProcessFlip requests that contains the
buffer addresses where the lock screen is being drawn to be
ignored, hence when resume happens the buffer being shown in
the screen is the incorrect one. This issue is very noticeable
when a bypass buffer is being used.

This patch fixes the previous issue by letting OMAPLFB to change
the overlay information until the DSS suspends the display even
if there is no composition present in the queue, so when the
system resumes the correct buffer address will be already
programmed in the DSS.

As part of this fix the omaplfb_drop_frame function is removed
since it was useful when DSSComp was managing the compositions
in the past.

Change-Id: I63f004d8a31cb3956ac9fa74dd8872d4a463db9f
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>