DSSCOMP: 3PDD: Fix tearing on auto update panels
Previously to present a new buffer in the display a call to manager->apply
and manager->wait_for_vsync was done. However there is no guarantee that
the new buffer address will be programmed for the next display refresh if
the manager->apply is done between the VFP and the VSYNC window. If this
happens the new buffer address will be held in the manager cache, leading
the client to think the buffer is already shown in the display if it waits
for one VSYNC unconditionally, tearing will be shown if the client writes
in the previously programmed buffer.

Calling wait_for_go instead of wait_for_vsync, makes sure the changes done in
the manager cache are in sync with the DSS shadow registers by waiting for the
necessary VSYNC periods for the changes to be reflected in the HW.

Change-Id: I39edd86de107d3b28b3a16a8cb6c867e41b18fea
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>