ASoC: McPDM: Fix the interrupt handler
Currently the McPDM interrupt handler is checking the MCPDM_IRQ_STATUS
register with irq. The status register needs to be checked against
irq_status which is the value read from it.

Fix the interrupt handler to check the status bits with the value
read from MCPDM_IRQ_STATUS register. Also for the info flags, change
from dev_err() to dev_dbg() to reduce the noise.

Change-Id: Id16ddfc57b90faa0d2c46d4522ae742e90a24499
Signed-off-by: Mohan V <mohanv@ti.com>
Cc: Misael Lopez Cruz <misael.lopez@ti.com>
Cc: Peter Ujfalusi <peter.ujfalusi@ti.com>