mmc:sd: Do not power on the regulator if SD plugged off
This is a workaround for prevent powering on SD card
if it is plugged off.

During SD card insertion/removing linux api is reading
some data from card for ensure the card is plugged off.
During this operation SD slot is powered up and dependently
from current PM state is powering off immidiately, after ~10s
or newer.

This patch prevents powering on SD card if it is plugged off.

Change-Id: I36f14d99f8ab32d4d35c38daf9e5b7df767108c9
Signed-off-by: Ruslan Bilovol <x0155540@ti.com>