OMAP4: DSI: Fixed incorrect frame in suspend-resume/warm-boot cases.
Blaze show a wrong frame in warm reboot or suspend/resume cases. This happens
because the blaze panel has framebuffer memory inside and that memory does
not cleared in suspend/resume or warm reboot cases. We can see a wrong frame
as result.
The patch prevents a displaying of panel memory while driver initialization.
Panel memory will be showed when DISPC update a screen.

Change-Id: Icf65fc2d59f6ed29ed1ac231b92bcc6a632ad7d8
Signed-off-by: Artem Rudenko <x0150883@ti.com>