SGX-KM: Row-by-row cache operation optimization
In case of a rectangular region of stride less than the screen
stride, we need to perform a cache operation on each row. This
was being done in a loop in the user side and involved an ioctl
for each row. This patch, along with corresponding user mode
changes, moves the loop to kernel side to perform the operation
using a single ioctl.

Change-Id: I15c1630aca93fc7fe9353708c009dd8eb48ab13a
Signed-off-by: Hemant Hariyani <hemanthariyani@ti.com>