usb: gadget: storage: remove debug warning from ep0 handler
During plugging blaze to host with WinXP while vendor request
is not processed composite_setup call all device registered
configure setup callbacks to process it.
This is used to handle the following cases:
 * - vendor request is for the device and arrives before setconfiguration.
 * - Some devices are required to handle vendor request before
     setconfiguration such as MTP, USBNET.

In this case fsg structures are not yet configured properly and
lot of debug output prints. This causes glitches in AV playback.

Replace *check with warnings print* by *check only*

Change-Id: I1b98a7a769d62f0636a877b0a5f5bf2bedb34e23
Signed-off-by: Ruslan Bilovol <x0155540@ti.com>