ARM: OMAP4: hwmod data: Change interface clocks for DSS hwmods
Change interface slave clocks of DSS hwmods from "dss_fck" to "l3_div_ck", this
would remove the need of having a fake clock called "dss_fck" to represent the
MODULEMODE bits of DSS clock domain.

Add the modulemode flag for "dss_core" hwmod so that MODULEMODE is enabled when
the hwmod is enabled. This works as DSS2 driver ensures that the platform device
corresponding to "dss_core" is the parent of all the other DSS hwmods and is
hence enabled first.

At bootup, all DSS hwmods apart from "dss_core" can't get enabled because they
don't independently enable MODULEMODE bits. The impact of this is not totally
clear.

Signed-off-by: Archit Taneja <archit@ti.com>