ASoC: davinci-mcasp: Add hw_rule for buffer_size when using AFIFO
The AFIFO threshold imposes a limiation on the buffer size.
When the AFIFO is used, the buffer size (in samples) needs to be
an integer multiple of the AFIFO threshold value (wnumevt, rnumevt).

This patch adds a hw_rule to the McASP driver for version 4
of the McASP to account for the limitation on the buffer size.

Change-Id: I3bd320130b10a55d3d84defd99572526ad6469fb
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Charulatha Varadarajan <charu@ti.com>