ASoC: davinci-mcasp: Pass pin inactive state via DT
The state of transmit/receive pins during inactive slots can be controlled
through DISMOD. Previously, the default state (Hi-Z) was used all the time, but
that mightvnot fit well for all applications. So, the inactive state is passed
via DT: "tx-inactive-state" and "rx-inactive-state".

The inactive states can actually be set per serializer, but for the sake of
simplicity we only differentiate them by stream direction.

Additionally, the DISMOD macro is also fixed as the argument part was incorrect.

Change-Id: I0e099c29fde94dd59175c1017c098554c43d18fe
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
Signed-off-by: Charulatha Varadarajan <charu@ti.com>