ASoC: DRA7: dra7-evm: gpio for DAI link pins switch
dra7evm board has switch/exchanger controlled through a gpio
that connects the pins of the Media and Bluetooth DAI links.
The DAI link pins (McASP2, McASP3 and McASP7) are connected
when the select gpio is low.

DAI sel gpio pin is driven by the pcf8575 which is on i2c2
bus. This bus shares pins with HDMI in DRA7 EVM, so the
DAI sel gpio request fails if HDMI is using the shared pins.

The default state for the gpio pin in the pcf8575 sets the
DAI sel switch to the correct state, so the gpio request
error is treated just as a warning.

Change-Id: Ie62b2082fdfa97828ee837daddaa527d56010e63
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Charulatha Varadarajan <charu@ti.com>
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>