VIP: Add an early sensor detect api
Sensor detection is required for early camera
as certain key resources such as pipeline and
the VIP are kept reserved for the early camera
use-case. In case the sensor is not present on
the board, we will have to detect that and bail
out without acquiring these resources. This patch
adds a sensor detection api for early camera
driver to query the presence of the sensor. It
does this by simply checking if the subdev
device pointer is populated in the device
structure and returning a boolean value to the
caller.

Change-Id: I2228517f0001decb486b5fa99f72d15428aad97e
Signed-off-by: Sundar Raman <sunds@ti.com>
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>