VIP: Early-Cam: Resolves empty buffer issue
During bootup, the early camera driver was
competing with Android to open up an instance
of the VIP module. This patch eliminates that
race condition. It also adds a delay in the
vip_buf_prepare function in order for the buffers
to be populated on rapid triggering of the GPIO
interrupt pin.

Change-Id: Ib1a4dc8e9b4242b00dcb035ba1712197be132943
Signed-off-by: Arthur Philpott <arthur.philpott@ti.com>
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>