omapdrm: Modify sync for write flags
omapdrm uses omap_gem_op_async to make sure the buffer is onyl accessed when
completely written by the SGX. The OMAP_GEM_READ flag makes it only read the
'read flag' in the buffer sync object. However, the SGX driver only updates the
'write flag' in the sync object. Add the OMAP_GEM_WRITE flag so that it
considers the 'write flag' in the sync object struct too.

Change-Id: Id8fc4f8ad06dbbe0632f77a444cd3f788398baee
Signed-off-by: Archit Taneja <archit@ti.com>