ARM: DRA7: hwmod_data: correct main clk for IPU1 MMU
The IPU1 functional clock is currently misconfigured to be
sourced from dpll_core_h22x2_ck. While this is true for IPU2,
there is a mux clock in between dpll_core_h22x2_ck and the
IPU1_GFCLK. Correct the hwmod data for the IPU1 MMU accordingly.
The mux clock is expected to be configured to be sourced from
the dpll_core_h22x2_ck in turn, so that both IPU1 and IPU2
are running at the same clock frequency.

Signed-off-by: Suman Anna <s-anna@ti.com>