ARM: DRA7: hwmod_data: correct main clk for IPU1
The IPU1 functional clock is currently misconfigured to have the
dpll_core_h22x2_mux clock as the main clock, but it is actually
the output of a mux clock, ipu1_gfclk_mux. Correct the hwmod data
appropriately to use this mux clock as the main clock. The mux clock
in turn will be configured to be sourced from the dpll_core_h22x2_ck,
and this ensures both IPU1 and IPU2 to be running from the same
clock source and at the same clock frequency.

Change-Id: I9d80eefcd7e71462d0e73bdc5704d586004140b2
Signed-off-by: Suman Anna <s-anna@ti.com>