ARM: DRA7: hwmod data: add bus-error back flag to ipu iommu attrs
The bus-error back response flag information has been added to
the iommu device attributes for IPU processor on DRA7. This raises
a bus error to the IPU processor, which can give out additional
stack trace of the MMU fault.

NOTE: This register is defined for the DSP MMUs as well, but
enabling it for DSPs suppresses the MMU fault address being
reported properly in the MMU_FAULT_AD register. The MMU_FAULT_PC
and MMU_FAULT_STATUS registers are not functional already on the
new C66 DSPs in DRA7.

Change-Id: Ie9be8dcd9979bf94121e5f2351fac4d6a2b316a6
Signed-off-by: Suman Anna <s-anna@ti.com>