rpmsg: rpc: cleanup the dma-buf idr completely at filp release
A complete cleanup of objects stored in an idr tree will require an
idr_destroy() to remove all id mappings, and to free all the cached
idr_layers, in addition to the freeing of all the objects using an
idr_for_each() loop. The idr_layers are internal structures that can
be allocated and cached within the idr library during the life cycle
of an idr tree.

The current dma-buf idr cleanup has been cleaning up only the
regular objects, and may end up having a memory leak of the cached
idr_layers. Cleanup this dma-buf idr completely when closing the
file handle.

Change-Id: I4cb5b5ed7df71d7a329cbb468c5b379377f0e92b
Reported-by: Subhajit Paul <subhajit_paul@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>