iommu/omap: add a index field for each iommu
Add an index field in omap_iommu object to store the index
of the IOMMU instance within a processor subsystem. The index
allows to identify each IOMMU instance in a processor subsystem
having multiple IOMMUs. The index 0 will be used by the default
MMU that a processor core goes through. The index is expected
to be supplied through the corresponding platform data.

This is in preparation to enable all the MMUs within DRA7xx
DSP sub-systems. DRA7xx is the first SoC in the OMAP family
to have more than a single MMU within a processor sub-system.

Change-Id: I8339c9e01f3f88373772ae260141c0c7fcb5f6c9
Signed-off-by: Suman Anna <s-anna@ti.com>