ARM: OMAP2+: remoteproc: split iommu archdata information
A remoteproc device is tied with its IOMMU through it's device
archdata field. The way this information is assigned currently
uses an array of iommu archdata elements, with each element
representing an IOMMU corresponding to a processor.

Split this iommu archdata array into discrete declarations to
prepare for supporting multiple iommus per processor. This info
could have been filled in directly in the respective platform
device declarations, but the discrete declarations keep both
the platform device and iommu archdata declarations to a minimum
with better code readability. The new iommu archdata is integrated
by introducing a new element to the struct omap_rproc_pdev_data.

Change-Id: I05d7133e38c567949a3d482a2a8f5553b8edc583
Signed-off-by: Suman Anna <s-anna@ti.com>