iommu/omap: change the attach detection logic
The OMAP IOMMU driver allows only a single device (eg: a rproc
device) to be attached per domain. The current detection logic
relies on an attached iommu for the client device. Change this
logic to use dedicated flags in preparation for supporting
multiple iommu devices to be bound to a single iommu domain.

Signed-off-by: Suman Anna <s-anna@ti.com>