iommu/omap: change the attach detection logic
The OMAP IOMMU driver allows only a single device (eg: a rproc
device) to be attached per domain. The current detection logic
relies on an attached iommu for the client device. Change this
logic to use dedicated flags in preparation for supporting
multiple iommu devices to be bound to a single iommu domain.

Change-Id: I30ef88a47e0883fd9d8369b1b337031768293569
Signed-off-by: Suman Anna <s-anna@ti.com>