ARM: OMAP2+: remoteproc: use sentinel terminated iommu archdata
A remoteproc device is tied with a single IOMMU at the moment,
and both are bound together through the rproc device's archdata
field, which is defined as a single element at present.

Change this iommu archdata information to be represented through
a sentinel terminated array, so that more than one MMU can be
added to a rproc device. The sentinel allows the IOMMU driver
core to be adapted to recognize multiple IOMMUs without having
to change the iommu archdata structure.

This logic scales well with both the current single IOMMU
instance per processor as well as with the upcoming enhanced
support in the OMAP IOMMU driver to program multiple IOMMUs
together.

Change-Id: I8d8ef0e85c024b4114fbbc5f95fbbd42161385e9
Signed-off-by: Suman Anna <s-anna@ti.com>