ARM: DRA7: hwmod data: remove resets from MMU1 for DSPs
The second MMU in the DSP processor subsystems, intended for
the providing address translations to the EDMA port, is also
controlled by the same reset as the first MMU, that provides
the translations to the processor cores.

This second MMU will not be programmed independently, it will
be programmed along-side the first MMU as a simple mirror copy.
The reset information need not be duplicated in both the MMUs,
as the resets will be ensured to be programmed properly while
configuring the first MMU. The reset data has therefore been
cleaned up. This also fixes a reset failure error in the hwmod
layer caused by the reset already being deasserted when the
data is not removed.

Change-Id: I9bb3e4825ee5389ad23576a43cd37d3d18a00bab
Signed-off-by: Suman Anna <s-anna@ti.com>