i2c: fpdlink: camera: Add fpdlink camera driver
It is possible to use output of FPDLink display as a camera
if it is connected to the deserializer of JAMR3 board.
Deserializer output is given to Vin1a

As, VIP can communicate with only subdevices, we need a subdevice to
use the fpdlink display as camera.

This patch adds a device tree node for the camera and the compatible driver.
The i2c driver for the fpdlink camera only registers a v4l2 subdevice
The display can be looped back from the same EVM or can come externally
via the FPDLink. In any case, deserializer configuration is done at the
remote end (serializer side), so this driver won't configure the deserializer.

Change-Id: I910aff26ff31444c38afda8fe5acd59ee2d8ec8f
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>