ARM: OMAP2+: Reserve memory for CMEM pool
In DRA7XX, radio application uses CMEM pool allocator for large
buffer sharing between MPU and DSP.

This patch does a static carveout for CMEM memory
area and sets aside this memory from the kernel. This needs to
be done early enough to avoid any ARM aliasing issues due to different
memory attributes that may be given to memory associated with the
kernel. This memory can be carved-out using bootargs, but this requires
every developer/user to be aware of this.

memblock API is used during the board initialization to reserve the
necessary memory.

In DRA7XX, as the CMEM reserved memory is shared between ARM and DSP
for Radio application,  the memory is reserved only if SOC_DRA7XX and
OMAP_REMOTEPROC_DSP config is defined.

Change-Id: I258b26e2b7fcdb0d39d500a0423c1e466852ba34
Tested-by: Tivy Robert <rtivy@ti.com>
Signed-off-by: Charulatha Varadarajan <charu@ti.com>