thermal: re-work poweroff strategy when critical temp is hit
The software controlled thermal driver, monitors for
cirital temperatures and triggers an orderly_poweroff() if
critical temperature is hit. This approach poses 2 problems:

1. orderly_poweroff() lets the userspace handle the poweroff
scenario before trying to do an emergency shutdown. Reaching
critical temperature calls for emergency shutdown. It is not
a good idea for the kernel to wait for the userspace to shutdown
the system.

2. The orderly_poweroff() implementation itself might be
asynchronous (the poweroff routine might be scheduled in a
workqueue), and thats not a good thing given the severity
of the event.

Hence trigger a kernel_poweroff() instead of an
orderly_poweroff().

However, there are platforms that cannot poweroff the system
(platform limitation, for eg: dra7xx where the PMIC cannot
be powered off). Hence as a last resort, provide a fallback
to restart the platform in case kernel_poweroff() is
not supported.

This patch takes an approach similar to the one present in the
emif driver introduced by Nishanth Menon <nm@ti.com> with the
commit:
cf653bbd9b346432a47373dd0fefd239ed34ad17

Change-Id: Ie637893f8298a66c892394411f8c15d44cff4e7f
Suggested-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Cc: Eduardo Valentin <eduardo.valentin@ti.com>