gadget: build fixes to get android gadget functional
The gadget framework and hence functions support in it have
migrated to CONFIGFS framework. This patch introduces changes
in the android gadget file to have it working with the new interfaces
for FunctionFS and Mass storage functions.

Also changed the Rndis functions to support older APIs to avoid build errors.

Signed-off-by: John Stultz <john.stultz@linaro.org>
[hashcode0f@gmail.com: removed changes to f_fs.c and reimplemented functionfs
as instance-based similar to mass_storage.  Fix compiler warning by adding
missing unbind_config function to mass_storage_functio]
[subramaniam.ca@ti.com: removed ffs_unbind_config. This causes the functions
registered to be removed multiple times causing a kernel panic. The functions
are already removed when usb_remove_config is called for the android gadget.
Also re-worded the commit description]

Change-Id: I6fe84f90789e4a4894ee6223f2bfb15639cd3e7b
Signed-off-by: Michael Scott <hashcode0f@gmail.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>