drm/omap: fix missing disable for unused encoder
When an encoder is no longer connected to a crtc, the driver will leave
the encoder enabled.

This patch adds code to track the encoder used for a crtc, and when the
encoder changes, the old one is disabled.

This patch with commit-id c7aef12f344459961eb1e0ba10d184816ed42d99 is i
backported from mainline.

Change-Id: Ifd929fdeea392a798dfb7d39573e0beb0bf9898f
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Reviewed-by: Rob Clark <robdclark@gmail.com>