i2c: tvp5158: Register subdevice only when analog camera detected
Registering a subdevice would mean that the camera is present
and the v4l2_device can use it and call subdev callbacks.
If the priority for the TVP camera is higher and no camera is connected,
Still it would be selected for capture resulting in a failure

This patch registers the subdevice only when the camera is detected and
signal is present.

Change-Id: I74ba9e399bf01a246fe7bc44f31482d203fd5c22
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>