v4l: ti-vps: vip: Set max width height as capture size
When an outbound descriptor is created, max width and height for
VPDMA are set to 1920x1080
When cameras give out data more than the size allocated, it may lead
to buffer corruption.

This patch uses different max width & height sets for different slices
and overrides the descriptor with width & height same as capture size.

Change-Id: I55b8c3098abb8841d4aba26e01a30ff612424f20
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>