v4l: ti-vps: vip: Handle short frames irqs
At any point of time, lists for FIFO_SIZE number of buffers are posted.
And driver expects those many distinct IRQs.

If a short frame comes, DMA for the frame would be completed immediately.
If VPDMA generates IRQ for short frame even before the IRQ for last frame
was cleared, both IRQs can be combined into one.

Once an IRQ is missed, number of buffers posted reduces by one.
This would lead to FIFO being empty, resulting in frame drop.

A better appraoch would be to post only one list in the streamON
In the IRQ handler, post for next buffer.
Do not dequeue the buffer only for first two IRQs

With this change, at any point of time, driver is posting only one list
So there is no chance of combining IRQs and leading to deadlock.

If a short frame occurs, only problem is that the next buffer would be missed
But capture continues as normal.

Change-Id: If5c8f09bc191fbe548543f7422740172a8d9e6a1
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>