drm/omap: Use old_fb to synchronize between successive page flips
omap_crtc->old_fb is used to check whether the previous page flip has completed
or not. However, it's never initialized to anything, so it's always NULL. This
results in the check to always succeed, and the page_flip to proceed.

Initialize old_fb to the fb that we intend to flip to through page_flip, and
therefore prevent a future page flip to proceed if the last one didn't
complete.

This patch with commit-id bc905aced30e48a39af7c452bf46228d7c6188b9 is backported
from mainline.

Conflicts:
	drivers/gpu/drm/omapdrm/omap_crtc.c

Change-Id: I1a73d6370d9594de4cfc15426b66bb809d31e760
Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Somnath Mukherjee <somnath@ti.com>