drm : avoid spinlock recursion by queueing notify
The notify function might hold spinlock which
will lead to recursion as sync_op_update is always
called with spinlock held.

The notify function can be called in from a worker
thread which is not an atomic context and thus
prevents spinlock recursion

Also, in case omap_gem_op_sync is interrupted
while it waits for notify to finish, two issues occur:
* the waiter struct is not freed
* the wait condition is on stack, so on return of the
  waiting function, the stack is corrupted
Therefore, the wait is made non-interruptible. So, when
omap_gem_op_sync returns, it is gurranteed that the gem
operation is complete.

Change-Id: I5a4350f4744173104a5a95ff2c92a829110047ff
Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>