v4l: ti-vps: vpe: use line average de-interlacing for first 2 frames
For n input fields, the VPE de-interlacer creates n - 2 progressive frames.

To support this, we use line average mode of de-interlacer for the first 2
input fields to generate 2 progressive frames. We then revert back to the
preferred EDI method, and create n - 2 frames, creating a sum of n frames.

V2:

- Fix issue when progressive frames run.
- Fix issue related to incorrect condition of video buffer removal.

Change-Id: I52c0642742086d7f65460ca23812ac99cdea1a99
Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>