usb: xhci: Add OTG support to xhci driver
This adds OTG support to the xhci driver, specifically for the
Synopsys PCI controller, but the code should be reusable for
other controllers.

The Synopsys controller requires disabling the host mode driver when
switching to peripheral mode, then re-enabling it when switching
back to host mode. We do that by calling usb_remove_hcd() /
usb_add_hcd().

This requires a small change to usb_hcd_pci_remove(), to test
whether usb_remove_hcd() has already been called and so should not
be called again when the driver is unloaded. We use the
hcd->rh_pollable flag for this

Change-Id: If1e4fcdec310976df41474840e8f00e614808300
Signed-off-by: Paul Zimmerman <Paul.Zimmerman@synopsys.com>
Signed-off-by: Ravi Babu <ravibabu@ti.com>