ARM: OMAP: remoteproc: add watchdog timer data for DRA7 IPU2
The watchdog timer information has been added to the device data
(timer data) associated with the IPU2 remoteproc device on DRA7 SoC.
The watchdog functionality will use GPTimers 4 and 9 for each of
the cores in IPU2. No watchdog timer is added to the remaining IPU
and DSP processors, but can be easily added as per need.

The BIOS-side code on IPU2 needs to configure/refresh this timer
properly to not throw a watchdog error.

Signed-off-by: Suman Anna <s-anna@ti.com>