ARM: DRA7: hwmod_data: remove modulemode from IPU/DSP hwmods
The .modulemode field is removed from all the IPU and DSP processor
hwmods. This fixes a kernel crash during the shutdown sequence of any
of the remoteproc devices, either during a normal teardown or during
a recovery.

The DSP and IPU processor subsystems are represented by multiple
hwmods - one for each of the MMUs present within the subsystem,
and one for the processor cores. The processor subsystem is clocked
from a single clock source with separate reset lines for each of
the processors and the MMUs. This clock and reset information is
represented in separate hwmods to allow the management of these
entities in different drivers operating on the corresponding
platform devices. This doesn't fit quite well into the current
omap_hwmod layer due to these inter-dependencies.

A remoteproc startup sequence involves enabling and programming of
the MMUs, loading of the firmware into RAM mapped by the MMUs, and
releasing the processors from reset. A shutdown sequence follows a
reverse pattern with the processors put into reset first followed by
the unmapping and disabling of the MMUs. Both the processors and the
MMUs are present within a single clock domain and requires the domain
be clocked and enabled until the last entity. The kernel crash is
caused during the unmapping phase of the MMUs, as the hwmod layer
disables the module during the omap_device_idle processing of the
processor hwmod. This issue is fixed by not defining a modulemode
for the IPU/DSP processors, which results in keeping the module in
an activated state. The module is disabled properly during the
omap_device_idle processing of the MMU hwmod while disabling the
MMU.

Signed-off-by: Suman Anna <s-anna@ti.com>