gc320: Fix hang during rmmod gccore
During removal of the module gccore there was a hang observed
due to am infinite wait at the command queue for the messages.

Normally, during termination of the gccore as in module exit,
a signal is sent to indicate to the command thread the thread
need to be shutdown or terminated.

The check for this signal was not at the appropriate place.
The command thread used to wait before checking if there
were any request for the termination.

This patch moves the check for termination / stop signals
before checking for any messages in the command queue.

Change-Id: I8d85303904ab3d4a8fe28e3a2fb680279b9c8ebb
Signed-off-by: Pradeep Venkatasubbarao <pradeepv@ti.com>