ARM: DRA7: Add wlan support for DRA7
The wilink module is a combo wireless connectivity sdio card based on
Texas Instrument's wl18xx solution.
It is a 4-wire, 1.8V, embedded sdio wlan device with an external irq
line and is power-controlled by a gpio-based fixed regulator.

Add support for the wlan capabilities of this module by muxing the
relevant mmc and gpio lines, and setting the required platform data.

Change-Id: Ia98db6868ea77b9fff4527ef03c71ba0229a3d76
Signed-off-by: Ido Yariv <ido@wizery.com>
Signed-off-by: bvijay <bvijay@ti.com>
Signed-off-by: Eyal Reizer <eyalr@ti.com>
Signed-off-by: Sriramakrishnan A G <srk@ti.com>