i2c: ov1063x: Fix kernel crash due to probe dependancy
In case of LVDS cameras, inside probe, driver configures the
corresponding deserializer and serializer devices first.
This will register the remote camera as a I2C alias to this I2C master.

Without this, any write to camera registers will fail. Therefore, there is
a dependancy on serializer/deserializer devices being probed before this
probed gets called. If that doesn't happen, driver tries to get the
I2C instance of the ser/des device from the DTS phandle
	(which would be NULL, since its not probed yet)
leading to a kernel crash.

Fix this by returning -EPROBE_DEFER in case the ser/des is not found.
This will make sure that the LVDS camera is probed only when the
serializer and deserializer devices are registered.

Added #ifndef check to avoid probde defer loop in case config isn't enabled.

Change-Id: Id5b80304b5775c6d2aade9d21084c9ff2e074d91
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>